Rename core/private.h to core/core_private.h
