+++ librz/core/core_private.h
